pyproject.toml
[project]
name = "streamlit-for-prototyping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.20.0",
    "google-generativeai>=0.8.5",
    "openai>=1.88.0",
    "streamlit>=1.46.0",
]
